Separate the basic building blocks in own module and add first complex model object for analytic function for mixing materials
