+++ orsopy/fileio/tests/test_model_language.py
@@ -11,28 +11,35 @@ from os.path import join as pjoin
 
 import pytest
 
+import orsopy.fileio.model_building_blocks
+
 from orsopy.fileio import ComplexValue, Value
+from orsopy.fileio import model_complex as mc
 from orsopy.fileio import model_language as ml
 
 
 class TestMaterial(unittest.TestCase):
     def test_empty(self):
         with self.assertRaises(ValueError):
-            mat = ml.Material()
+            mat = orsopy.fileio.model_building_blocks.Material()
             mat.resolve_defaults({})
 
     def test_values(self):
-        m = ml.Material(formula="Fe2O3", mass_density={"magnitude": 7.0, "unit": "g/cm^3"})
+        m = orsopy.fileio.model_building_blocks.Material(
+            formula="Fe2O3", mass_density={"magnitude": 7.0, "unit": "g/cm^3"}
+        )
         assert m.mass_density == Value(7.0, "g/cm^3")
 
     def test_default(self):
-        defaults = ml.ModelParameters(
+        defaults = orsopy.fileio.model_building_blocks.ModelParameters(
             mass_density_unit="g/cm^3",
             number_density_unit="1/nm^3",
             sld_unit="1/angstrom^2",
             magnetic_moment_unit="muB",
         )
-        m = ml.Material(formula="Fe2O3", mass_density=7.0, number_density=0.15, sld=6.3e-6, magnetic_moment=3.4)
+        m = orsopy.fileio.model_building_blocks.Material(
+            formula="Fe2O3", mass_density=7.0, number_density=0.15, sld=6.3e-6, magnetic_moment=3.4
+        )
         m.resolve_defaults(defaults)
 
         assert m.mass_density == Value(7.0, "g/cm^3")
@@ -40,7 +47,7 @@ class TestMaterial(unittest.TestCase):
         assert m.sld == Value(6.3e-6, "1/angstrom^2")
         assert m.magnetic_moment == Value(3.4, "muB")
 
-        m = ml.Material(
+        m = orsopy.fileio.model_building_blocks.Material(
             formula="Fe2O3",
             mass_density=Value(7.0),
             number_density=Value(0.15),
@@ -51,43 +58,43 @@ class TestMaterial(unittest.TestCase):
 
     def test_density_lookup_elements(self):
         # no lookup case
-        m = ml.Material(sld=Value(4e-6, "1/angstrom^3"))
+        m = orsopy.fileio.model_building_blocks.Material(sld=Value(4e-6, "1/angstrom^3"))
         m.generate_density()
         # single element lookup case
         for element in ["Co", "Ni", "Si", "C"]:
-            m = ml.Material(formula=element)
+            m = orsopy.fileio.model_building_blocks.Material(formula=element)
             m.generate_density()
             # repeat for caching
-            m = ml.Material(formula=element)
+            m = orsopy.fileio.model_building_blocks.Material(formula=element)
             m.generate_density()
             assert m.number_density is not None
             assert m.comment.startswith("density from ORSO SLD db ID")
         # mixed element lookup case
         for element in ["Co0.8Cr0.2Fe0.1", "Ni3.4O5.4", "SiC4.3425"]:
-            m = ml.Material(formula=element)
+            m = orsopy.fileio.model_building_blocks.Material(formula=element)
             m.generate_density()
             assert m.number_density is not None
             assert m.comment == "density from average element density from ORSO SLD db"
 
     def test_sld(self):
-        m = ml.Material(sld=ComplexValue(3.4e-6, -2e-6, "1/angstrom^2"))
+        m = orsopy.fileio.model_building_blocks.Material(sld=ComplexValue(3.4e-6, -2e-6, "1/angstrom^2"))
         assert m.get_sld() == (3.4e-6 - 2e-6j)
-        m = ml.Material(formula="Si", mass_density=Value(2.33, "g/cm^3"))
+        m = orsopy.fileio.model_building_blocks.Material(formula="Si", mass_density=Value(2.33, "g/cm^3"))
         self.assertAlmostEqual(m.get_sld().real, 2.07371e-6, 5)
         self.assertAlmostEqual(m.get_sld().imag, -0.00002e-6, 5)
-        m = ml.Material(formula="Si", number_density=Value(0.04996026, "1/angstrom^3"))
+        m = orsopy.fileio.model_building_blocks.Material(formula="Si", number_density=Value(0.04996026, "1/angstrom^3"))
         self.assertAlmostEqual(m.get_sld().real, 2.07371e-6, 5)
         self.assertAlmostEqual(m.get_sld().imag, -0.00002e-6, 5)
-        m = ml.Material(formula="Si", number_density=Value(49.96026, "1/nm^3"))
+        m = orsopy.fileio.model_building_blocks.Material(formula="Si", number_density=Value(49.96026, "1/nm^3"))
         self.assertAlmostEqual(m.get_sld().real, 2.07371e-6, 5)
         self.assertAlmostEqual(m.get_sld().imag, -0.00002e-6, 5)
-        m = ml.Material(formula="Si")
+        m = orsopy.fileio.model_building_blocks.Material(formula="Si")
         assert m.get_sld() == (0.0j)
 
     def test_to_yaml(self):
-        m = ml.Material(sld=ComplexValue(3.4e-6, -2e-6, "1/angstrom^2"))
+        m = orsopy.fileio.model_building_blocks.Material(sld=ComplexValue(3.4e-6, -2e-6, "1/angstrom^2"))
         assert m.to_yaml() == "sld: {real: 3.4e-06, imag: -2.0e-06, unit: 1/angstrom^2}\n"
-        m = ml.Material(
+        m = orsopy.fileio.model_building_blocks.Material(
             formula="Si",
             mass_density=Value(2.33, "g/cm^3"),
             number_density=Value(0.04996026, "1/angstrom^3"),
@@ -104,26 +111,28 @@ class TestMaterial(unittest.TestCase):
 class TestComposit(unittest.TestCase):
     def test_wrong_value(self):
         with pytest.warns(RuntimeWarning):
-            ml.Composit(composition="123")
+            orsopy.fileio.model_building_blocks.Composit(composition="123")
 
     def test_resolution(self):
-        materials = {"Si": ml.Material(formula="Si")}
-        c = ml.Composit({"air": 0.3, "water": 0.3, "Si": 0.2, "Co": 0.1})
+        materials = {"Si": orsopy.fileio.model_building_blocks.Material(formula="Si")}
+        c = orsopy.fileio.model_building_blocks.Composit({"air": 0.3, "water": 0.3, "Si": 0.2, "Co": 0.1})
         c.resolve_names(materials)
         for key in ["air", "water", "Si", "Co"]:
             assert key in c._composition_materials
 
     def test_defaults(self):
-        defaults = ml.ModelParameters(
+        defaults = orsopy.fileio.model_building_blocks.ModelParameters(
             mass_density_unit="g/cm^3",
             number_density_unit="1/nm^3",
             sld_unit="1/angstrom^2",
             magnetic_moment_unit="muB",
         )
         materials = {
-            "Si": ml.Material(formula="Fe2O3", mass_density=7.0, number_density=0.15, sld=6.3e-6, magnetic_moment=3.4)
+            "Si": orsopy.fileio.model_building_blocks.Material(
+                formula="Fe2O3", mass_density=7.0, number_density=0.15, sld=6.3e-6, magnetic_moment=3.4
+            )
         }
-        c = ml.Composit({"Si": 1.0})
+        c = orsopy.fileio.model_building_blocks.Composit({"Si": 1.0})
         c.resolve_names(materials)
         c.resolve_defaults(defaults)
         m = materials["Si"]
@@ -133,15 +142,15 @@ class TestComposit(unittest.TestCase):
         assert m.magnetic_moment == Value(3.4, "muB")
 
     def test_density_lookup_elements(self):
-        materials = {"Si": ml.Material(sld=Value(2.0e-6, "1/angstrom^2"))}
-        c = ml.Composit({"Si": 0.5})
+        materials = {"Si": orsopy.fileio.model_building_blocks.Material(sld=Value(2.0e-6, "1/angstrom^2"))}
+        c = orsopy.fileio.model_building_blocks.Composit({"Si": 0.5})
         c.resolve_names(materials)
         c.generate_density()
 
         self.assertAlmostEqual(c.get_sld(), 1.0e-6)
 
     def test_to_yaml(self):
-        c = ml.Composit({"air": 0.3, "water": 0.3, "Si": 0.2, "Co": 0.1})
+        c = orsopy.fileio.model_building_blocks.Composit({"air": 0.3, "water": 0.3, "Si": 0.2, "Co": 0.1})
         assert c.to_yaml() == "composition:\n  air: 0.3\n  water: 0.3\n  Si: 0.2\n  Co: 0.1\n"
 
 
@@ -154,79 +163,85 @@ class TestLayer(unittest.TestCase):
 
     def test_empty(self):
         with self.assertRaises(ValueError):
-            lay = ml.Layer()
+            lay = orsopy.fileio.model_building_blocks.Layer()
             lay.resolve_names({})
 
     def test_resolution(self):
-        materials = {"Si": ml.Material(formula="Si")}
-        lay = ml.Layer(composition={"air": 0.3, "water": 0.3, "Si": 0.2, "Co": 0.1})
+        materials = {"Si": orsopy.fileio.model_building_blocks.Material(formula="Si")}
+        lay = orsopy.fileio.model_building_blocks.Layer(composition={"air": 0.3, "water": 0.3, "Si": 0.2, "Co": 0.1})
         lay.resolve_names(materials)
         for key in ["air", "water", "Si", "Co"]:
             assert key in lay._composition_materials
 
-        lay = ml.Layer(material="Si")
+        lay = orsopy.fileio.model_building_blocks.Layer(material="Si")
         lay.resolve_names(materials)
         assert lay.material is materials["Si"]
 
-        lay = ml.Layer(material="air")
+        lay = orsopy.fileio.model_building_blocks.Layer(material="air")
         lay.resolve_names(materials)
-        assert lay.material is ml.SPECIAL_MATERIALS["air"]
+        assert lay.material is orsopy.fileio.model_building_blocks.SPECIAL_MATERIALS["air"]
 
-        lay = ml.Layer(material=materials["Si"])
+        lay = orsopy.fileio.model_building_blocks.Layer(material=materials["Si"])
         lay.resolve_names(materials)
         assert lay.material is materials["Si"]
 
-        c = ml.Composit({"Si": 1.0})
-        lay = ml.Layer(material=c)
+        c = orsopy.fileio.model_building_blocks.Composit({"Si": 1.0})
+        lay = orsopy.fileio.model_building_blocks.Layer(material=c)
         lay.resolve_names(materials)
         assert lay.material is c
 
-        lay = ml.Layer(material="Si")
+        lay = orsopy.fileio.model_building_blocks.Layer(material="Si")
         lay.resolve_names({})
         assert lay.material.formula == "Si"
 
     def test_defaults(self):
-        defaults = ml.ModelParameters(
+        defaults = orsopy.fileio.model_building_blocks.ModelParameters(
             mass_density_unit="g/cm^3",
             number_density_unit="1/nm^3",
             sld_unit="1/angstrom^2",
             magnetic_moment_unit="muB",
         )
 
-        lay = ml.Layer(material=ml.Material(sld=2.0e-6))
+        lay = orsopy.fileio.model_building_blocks.Layer(
+            material=orsopy.fileio.model_building_blocks.Material(sld=2.0e-6)
+        )
         lay.resolve_defaults(defaults)
 
         assert lay.thickness == Value(0.0, defaults.length_unit)
         assert lay.roughness == defaults.roughness
         assert lay.material.sld.unit == defaults.sld_unit
 
-        lay = ml.Layer(material=ml.Material(sld=Value(2.0e-6)), thickness=31.2, roughness=1.3)
+        lay = orsopy.fileio.model_building_blocks.Layer(
+            material=orsopy.fileio.model_building_blocks.Material(sld=Value(2.0e-6)), thickness=31.2, roughness=1.3
+        )
         lay.resolve_defaults(defaults)
 
         assert lay.thickness == Value(31.2, defaults.length_unit)
         assert lay.roughness == Value(1.3, defaults.length_unit)
         assert lay.material.sld.unit == defaults.sld_unit
 
-        lay = ml.Layer(
-            material=ml.Material(sld=Value(2.0e-6, defaults.sld_unit)), thickness=Value(31.2), roughness=Value(1.3)
+        lay = orsopy.fileio.model_building_blocks.Layer(
+            material=orsopy.fileio.model_building_blocks.Material(sld=Value(2.0e-6, defaults.sld_unit)),
+            thickness=Value(31.2),
+            roughness=Value(1.3),
         )
         lay.resolve_defaults(defaults)
 
         assert lay.thickness == Value(31.2, defaults.length_unit)
         assert lay.roughness == Value(1.3, defaults.length_unit)
 
-        lay = ml.Layer(composition={"Si": 1.0})
-        lay.resolve_names({"Si": ml.Material(sld=2.0e-6)})
+        lay = orsopy.fileio.model_building_blocks.Layer(composition={"Si": 1.0})
+        lay.resolve_names({"Si": orsopy.fileio.model_building_blocks.Material(sld=2.0e-6)})
         lay.resolve_defaults(defaults)
 
         assert lay._composition_materials["Si"].sld.unit == defaults.sld_unit
 
     def test_material_mixing(self):
-        lay = ml.Layer(composition={"Si": 1.0})
-        lay.resolve_names({"Si": ml.Material(sld=Value(2.0e-6, unit="1/angstrom^2"))})
+        lay = orsopy.fileio.model_building_blocks.Layer(composition={"Si": 1.0})
+        lay.resolve_names({"Si": orsopy.fileio.model_building_blocks.Material(sld=Value(2.0e-6, unit="1/angstrom^2"))})
         lay.generate_material()
 
-        assert isinstance(lay.material, ml.Material)
+        assert isinstance(lay.material, orsopy.fileio.model_building_blocks.Material)
         assert lay.material.sld == ComplexValue(2.0e-6, 0.0, unit="1/angstrom^2")
 
 
@@ -239,56 +254,87 @@ class TestSubStack(unittest.TestCase):
     def test_resolution(self):
         s = ml.SubStack(stack="air | b 13 |c|d")
         resolvable_items = {
-            "d": ml.Layer(material=ml.Material(sld=Value(2e-6, "1/angstrom^2"))),
-            "b": ml.Material(formula="Co"),
-            "c": ml.Composit({"b": 1.0}),
+            "d": orsopy.fileio.model_building_blocks.Layer(
+                material=orsopy.fileio.model_building_blocks.Material(sld=Value(2e-6, "1/angstrom^2"))
+            ),
+            "b": orsopy.fileio.model_building_blocks.Material(formula="Co"),
+            "c": orsopy.fileio.model_building_blocks.Composit({"b": 1.0}),
         }
         s.resolve_names(resolvable_items)
         assert len(s.sequence) == 4
-        assert s.sequence[0] == ml.Layer(thickness=0.0, material=ml.SPECIAL_MATERIALS["air"])
-        assert s.sequence[1] == ml.Layer(thickness=13.0, material=ml.Material(formula="Co"))
-        assert s.sequence[2] == ml.Layer(thickness=0.0, material=ml.Composit({"b": 1.0}))
+        assert s.sequence[0] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=0.0, material=orsopy.fileio.model_building_blocks.SPECIAL_MATERIALS["air"]
+        )
+        assert s.sequence[1] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=13.0, material=orsopy.fileio.model_building_blocks.Material(formula="Co")
+        )
+        assert s.sequence[2] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=0.0, material=orsopy.fileio.model_building_blocks.Composit({"b": 1.0})
+        )
         assert s.sequence[3] == resolvable_items["d"]
 
         s = ml.SubStack(stack="air | 2( b 13 | c 5)|d")
         s.resolve_names(resolvable_items)
         assert len(s.sequence) == 3
-        assert s.sequence[0] == ml.Layer(thickness=0.0, material=ml.SPECIAL_MATERIALS["air"])
+        assert s.sequence[0] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=0.0, material=orsopy.fileio.model_building_blocks.SPECIAL_MATERIALS["air"]
+        )
         assert s.sequence[1] == ml.SubStack(
             repetitions=2,
             stack="b 13 | c 5",
             sequence=[
-                ml.Layer(thickness=13.0, material=ml.Material(formula="Co")),
-                ml.Layer(thickness=5.0, material=ml.Composit(composition={"b": 1.0})),
+                orsopy.fileio.model_building_blocks.Layer(
+                    thickness=13.0, material=orsopy.fileio.model_building_blocks.Material(formula="Co")
+                ),
+                orsopy.fileio.model_building_blocks.Layer(
+                    thickness=5.0, material=orsopy.fileio.model_building_blocks.Composit(composition={"b": 1.0})
+                ),
             ],
         )
         assert s.sequence[2] == resolvable_items["d"]
 
-        s = ml.SubStack(sequence=[ml.Layer(thickness=13.0, material="b"), ml.Layer(thickness=5.0, material="c")])
+        s = ml.SubStack(
+            sequence=[
+                orsopy.fileio.model_building_blocks.Layer(thickness=13.0, material="b"),
+                orsopy.fileio.model_building_blocks.Layer(thickness=5.0, material="c"),
+            ]
+        )
         s.resolve_names(resolvable_items)
         assert s.sequence == [
-            ml.Layer(thickness=13.0, material=ml.Material(formula="Co")),
-            ml.Layer(thickness=5.0, material=ml.Composit(composition={"b": 1.0})),
+            orsopy.fileio.model_building_blocks.Layer(
+                thickness=13.0, material=orsopy.fileio.model_building_blocks.Material(formula="Co")
+            ),
+            orsopy.fileio.model_building_blocks.Layer(
+                thickness=5.0, material=orsopy.fileio.model_building_blocks.Composit(composition={"b": 1.0})
+            ),
         ]
 
     def test_defaults(self):
-        defaults = ml.ModelParameters(
+        defaults = orsopy.fileio.model_building_blocks.ModelParameters(
             mass_density_unit="g/cm^3",
             number_density_unit="1/nm^3",
             sld_unit="1/angstrom^2",
             magnetic_moment_unit="muB",
         )
 
-        s = ml.SubStack(sequence=[ml.Layer(thickness=13.0, material=ml.Material(formula="Co"))])
+        s = ml.SubStack(
+            sequence=[
+                orsopy.fileio.model_building_blocks.Layer(
+                    thickness=13.0, material=orsopy.fileio.model_building_blocks.Material(formula="Co")
+                )
+            ]
+        )
         s.resolve_defaults(defaults)
         assert s.sequence[0].thickness == Value(13.0, defaults.length_unit)
 
     def test_resolve_layers(self):
         resolvable_items = {
-            "d": ml.Layer(material=ml.Material(sld=Value(2e-6, "1/angstrom^2"))),
-            "b": ml.Material(formula="Co"),
-            "c": ml.Composit({"b": 1.0}),
-            "e": ml.Layer(composition={"b": 1.0}),
+            "d": orsopy.fileio.model_building_blocks.Layer(
+                material=orsopy.fileio.model_building_blocks.Material(sld=Value(2e-6, "1/angstrom^2"))
+            ),
+            "b": orsopy.fileio.model_building_blocks.Material(formula="Co"),
+            "c": orsopy.fileio.model_building_blocks.Composit({"b": 1.0}),
+            "e": orsopy.fileio.model_building_blocks.Layer(composition={"b": 1.0}),
         }
         s = ml.SubStack(stack="air | 2( b 13 | c 5)|d|e")
         s.resolve_names(resolvable_items)
@@ -299,9 +345,9 @@ class TestSubStack(unittest.TestCase):
 class TestSampleModel(unittest.TestCase):
     def test_resolvable_items(self):
         sub_stacks = {"a": ml.SubStack(stack="b")}
-        layers = {"b": ml.Layer(thickness=13.4, material="c")}
-        materials = {"c": ml.Material(sld=13.4)}
-        composits = {"d": ml.Composit({"c": 1.0})}
+        layers = {"b": orsopy.fileio.model_building_blocks.Layer(thickness=13.4, material="c")}
+        materials = {"c": orsopy.fileio.model_building_blocks.Material(sld=13.4)}
+        composits = {"d": orsopy.fileio.model_building_blocks.Composit({"c": 1.0})}
         sm = ml.SampleModel(
             stack="c|a|c", sub_stacks=sub_stacks, layers=layers, materials=materials, composits=composits
         )
@@ -313,21 +359,21 @@ class TestSampleModel(unittest.TestCase):
         with pytest.warns(UserWarning):
             ml.SampleModel(
                 stack="c|a|c",
-                layers={"a": ml.Layer(thickness=13.4, material="c")},
-                materials={"a": ml.Material(sld=13.4)},
+                layers={"a": orsopy.fileio.model_building_blocks.Layer(thickness=13.4, material="c")},
+                materials={"a": orsopy.fileio.model_building_blocks.Material(sld=13.4)},
             )
 
     def test_space_in_name(self):
         sm = ml.SampleModel(
             stack="c|a b|c",
-            layers={"a": ml.Layer(thickness=13.4, material="c")},
-            materials={"a b": ml.Material(sld=13.4)},
+            layers={"a": orsopy.fileio.model_building_blocks.Layer(thickness=13.4, material="c")},
+            materials={"a b": orsopy.fileio.model_building_blocks.Material(sld=13.4)},
         )
         res = sm.resolvable_items
         assert list(res.keys()) == ["a", "a b"]
 
     def test_resolve_stack(self):
-        defaults = ml.ModelParameters(
+        defaults = orsopy.fileio.model_building_blocks.ModelParameters(
             mass_density_unit="g/cm^3",
             number_density_unit="1/nm^3",
             sld_unit="1/angstrom^2",
@@ -337,9 +383,9 @@ class TestSampleModel(unittest.TestCase):
         )
 
         sub_stacks = {"a": ml.SubStack(stack="b")}
-        layers = {"b": ml.Layer(thickness=13.4, material="c")}
-        materials = {"c": ml.Material(sld=13.4)}
-        composits = {"d": ml.Composit({"c": 1.0})}
+        layers = {"b": orsopy.fileio.model_building_blocks.Layer(thickness=13.4, material="c")}
+        materials = {"c": orsopy.fileio.model_building_blocks.Material(sld=13.4)}
+        composits = {"d": orsopy.fileio.model_building_blocks.Composit({"c": 1.0})}
         sm = ml.SampleModel(
             stack="c|2(a|c 15)|d 14|c",
             sub_stacks=sub_stacks,
@@ -353,21 +399,23 @@ class TestSampleModel(unittest.TestCase):
         subs.resolve_names({"a": sub_stacks["a"], "c": materials["c"]})
         subs.resolve_defaults(defaults)
         assert len(stack) == 4
-        assert stack[0] == ml.Layer(
-            thickness=ml.Value(0.0, "nm"), roughness=defaults.roughness, material=materials["c"]
+        assert stack[0] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=Value(0.0, "nm"), roughness=defaults.roughness, material=materials["c"]
         )
         assert stack[1] == subs
-        assert stack[2] == ml.Layer(
-            thickness=ml.Value(14.0, "nm"), roughness=defaults.roughness, material=composits["d"]
+        assert stack[2] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=Value(14.0, "nm"), roughness=defaults.roughness, material=composits["d"]
         )
-        assert stack[3] == ml.Layer(
-            thickness=ml.Value(0.0, "nm"), roughness=defaults.roughness, material=materials["c"]
+        assert stack[3] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=Value(0.0, "nm"), roughness=defaults.roughness, material=materials["c"]
         )
         sm = ml.SampleModel("Si")
         stack = sm.resolve_stack()
         assert len(stack) == 1
-        assert stack[0] == ml.Layer(
-            ml.Value(0.0, "nm"), roughness=defaults.roughness, material=ml.Material(formula="Si")
+        assert stack[0] == orsopy.fileio.model_building_blocks.Layer(
+            Value(0.0, "nm"),
+            roughness=defaults.roughness,
+            material=orsopy.fileio.model_building_blocks.Material(formula="Si"),
         )
 
     def test_resolve_direct_name(self):
@@ -388,8 +436,23 @@ class TestSampleModel(unittest.TestCase):
         self.assertEqual(stack[1].material.formula, "D2O")
         self.assertEqual(stack[1].thickness.magnitude, 12.0)
 
+    def test_resovle_function2e(self):
+        sm = ml.SampleModel(
+            stack="air | gradient | Si",
+            sub_stacks={
+                "gradient": mc.FunctionTwoElements(
+                    material1="Cr", material2="Fe", total_thickness=150.0, function="x", slices=10
+                )
+            },
+        )
+        layers = sm.resolve_to_layers()
+        self.assertEqual(len(layers), 12)
+        for li in layers:
+            li.material.generate_density()
+            li.material.get_sld()
+
     def test_resolve_to_layers(self):
-        defaults = ml.ModelParameters(
+        defaults = orsopy.fileio.model_building_blocks.ModelParameters(
             mass_density_unit="g/cm^3",
             number_density_unit="1/nm^3",
             sld_unit="1/angstrom^2",
@@ -399,9 +462,12 @@ class TestSampleModel(unittest.TestCase):
         )
 
         sub_stacks = {"a": ml.SubStack(stack="b")}
-        layers = {"b": ml.Layer(thickness=13.4, material="c"), "b2": ml.Layer(thickness=2.0, composition={"c": 1.0})}
-        materials = {"c": ml.Material(sld=13.4)}
-        composits = {"d": ml.Composit({"c": 1.0})}
+        layers = {
+            "b": orsopy.fileio.model_building_blocks.Layer(thickness=13.4, material="c"),
+            "b2": orsopy.fileio.model_building_blocks.Layer(thickness=2.0, composition={"c": 1.0}),
+        }
+        materials = {"c": orsopy.fileio.model_building_blocks.Material(sld=13.4)}
+        composits = {"d": orsopy.fileio.model_building_blocks.Composit({"c": 1.0})}
         sm = ml.SampleModel(
             stack="c|2(a|c 15)|b|b2|d 14|Si",
             sub_stacks=sub_stacks,
@@ -410,19 +476,29 @@ class TestSampleModel(unittest.TestCase):
             composits=composits,
             globals=defaults,
         )
-        mSi = ml.Material(formula="Si")
+        mSi = orsopy.fileio.model_building_blocks.Material(formula="Si")
         mSi.generate_density()
         ls = sm.resolve_to_layers()
         assert len(ls) == 9
-        assert ls[0] == ml.Layer(thickness=ml.Value(0.0, "nm"), roughness=defaults.roughness, material=materials["c"])
+        assert ls[0] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=Value(0.0, "nm"), roughness=defaults.roughness, material=materials["c"]
+        )
         assert ls[1] == ls[3]
         assert ls[2] == ls[4]
-        assert ls[5] == ml.Layer(thickness=ml.Value(13.4, "nm"), roughness=defaults.roughness, material=materials["c"])
-        assert ls[6] == ml.Layer(
-            thickness=ml.Value(2.0, "nm"),
+        assert ls[5] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=Value(13.4, "nm"), roughness=defaults.roughness, material=materials["c"]
+        )
+        assert ls[6] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=Value(2.0, "nm"),
             roughness=defaults.roughness,
-            material=ml.Material(sld=ComplexValue(13.4, 0.0, "1/angstrom^2"), comment="composition material: 1.0xc"),
+            material=orsopy.fileio.model_building_blocks.Material(
+                sld=ComplexValue(13.4, 0.0, "1/angstrom^2"), comment="composition material: 1.0xc"
+            ),
             composition={"c": 1.0},
         )
-        assert ls[7] == ml.Layer(thickness=ml.Value(14.0, "nm"), roughness=defaults.roughness, material=composits["d"])
-        assert ls[8] == ml.Layer(thickness=Value(0.0, "nm"), roughness=defaults.roughness, material=mSi)
+        assert ls[7] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=Value(14.0, "nm"), roughness=defaults.roughness, material=composits["d"]
+        )
+        assert ls[8] == orsopy.fileio.model_building_blocks.Layer(
+            thickness=Value(0.0, "nm"), roughness=defaults.roughness, material=mSi
+        )
